Changed global & local functions to diff. files
v1.5.9
