+++ rtos/rt_task.h
@@ -12,9 +12,10 @@
 
 #define     max_active_TCB      32
 
-int OS_Task_Create(voidfuncptr task_entry, void *argv);
-void OS_Task_Delete(voidfuncptr task);
 void __empty(void);
+OS_TID rt_get_TID(void);
+P_TCB rt_tsk_create(voidfuncptr task_entry, void *argv);
+int rt_tsk_delete(OS_TID task_id);
 
 extern void *os_active_TCB[];
